requirements.txt
beautifulsoup4
dateparser
discord.py
openai
Pillow
python-dotenv
Requests
tabulate
validators
yt_dlp
PyNaCl
spotdl
javascript